groundstation/requirements.txt
Flask==2.2.2
pytest==7.1.3
influxdb-client==1.33.0
Flask-SocketIO==5.3.1
py~=1.11.0
python-dotenv~=0.21.0
flask_cors==3.0.10
simple-websocket==0.8.1
redis